Revert "Create classes (#11)"
This reverts commit 25f5babb7f73929c71b4204f4fc032fe5e297875.
